Use the official Ubuntu AMI.
This is the option listed on the "Quick start" tab of the AMI selector
in AWS. The previous AMI comes from an unknown source and contains an
unknown and weird entry in authorized_keys which was determined to have
come from the base AMI that was being used. Since the image in question
was of unknown provenance, it has been assumed to be compromised and
therefore replaced with a known good one.